+++ src/roiextractors/extractors/inscopixextractors/inscopiximagingextractor.py
@@ -66,10 +66,6 @@ class InscopixImagingExtractor(ImagingExtractor):
         warnings.warn("isx only supports single channel videos.")
         return ["channel_0"]
 
-    def get_num_channels(self) -> int:
-        warnings.warn("isx only supports single channel videos.")
-        return 1
-
     def get_series(self, start_sample: Optional[int] = None, end_sample: Optional[int] = None) -> np.ndarray:
         start_sample = start_sample or 0
         end_sample = end_sample or self.get_num_samples()