+++ src/roiextractors/extractors/nwbextractors/nwbextractors.py
@@ -308,9 +308,6 @@ class NwbImagingExtractor(ImagingExtractor):
     def get_channel_names(self):
         return self._channel_names
 
-    def get_num_channels(self):
-        return self._num_channels
-
     def get_native_timestamps(
         self, start_sample: Optional[int] = None, end_sample: Optional[int] = None
     ) -> Optional[np.ndarray]: