+++ tests/test_micromanagertiffimagingextractor.py
@@ -77,9 +77,6 @@ class TestMicroManagerTiffExtractor(TestCase):
     def test_micromanagertiffextractor_channel_names(self):
         self.assertEqual(self.extractor.get_channel_names(), ["Default"])
 
-    def test_micromanagertiffextractor_num_channels(self):
-        self.assertEqual(self.extractor.get_num_channels(), 1)
-
     def test_micromanagertiffextractor_dtype(self):
         self.assertEqual(self.extractor.get_dtype(), np.uint16)
 
@@ -92,9 +89,6 @@ class TestMicroManagerTiffExtractor(TestCase):
     def test_private_micromanagertiffextractor_num_frames(self):
         for sub_extractor in self.extractor._imaging_extractors:
             self.assertEqual(sub_extractor.get_num_frames(), 5)
-
-    def test_private_micromanagertiffextractor_num_channels(self):
-        self.assertEqual(self.extractor._imaging_extractors[0].get_num_channels(), 1)
 
     def test_private_micromanagertiffextractor_sampling_frequency(self):
         sub_extractor = self.extractor._imaging_extractors[0]